example.py: restore changes lost by bad merge
ea5e675 (update example.py, 2015-06-30) introduced a bunch of
improvements to example.py, and most of these changes were accidentally
reverted by 4907e63 (Merge branch 'Xuefeng-Zhu-example', 2015-11-24).

In particular, the original commit introduced the creation of a
mechanicalsoup.Form() object based on a beautiful soup one, and this
change is needed for the code to work. Without this, one tries to call
method input() on a soup object instead of using the mechanicalsoup one.

Fixes #40.