StatefulBrowser implementation
Fixes #48.

StatefulBrowser is an extension of Browser than remembers the current
page and URL, to reduce the amount of data the caller has to deal with.
It also allows selecting a form in the page, which allows a method
submit_selected() to do all the work to submit a form (pick the form and
the URL).

The code is slightly adapted from chamiloBrowser.py from
https://gitlab.com/chamilotools/chamilotools/. I'm the original author
of all the code and I hereby licence it under The MIT License (MIT), for
inclusion into MechanicalSoup.

The code has no proper test as part of MechanicalSoup yet, but it can
run the regression suite of chamilotools.