example.py: specify parser explicitly
It's really bad practice to let the system decide which parser to use:
it encourages people to write scripts that will behave differently on
different machines, depending on which parser is installed.

example.py does work with lxml, but not with html.parser for example,
hence the example is broken on machines where html.parser is available
but lxml is not.

Fix this by specifying the parser explicitly.

IMHO, this is the proper fix to the issue raised by PR #67.