Apply user_agent to all session requests
The previous implementation of user_agent only set the user_agent
for requests created with `Browser._build_request`. However, there
are many other methods in `Browser` that make requests, and these
did not use the specified user_agent.

The `test_user_agent` function is added to show the failure
scenario. To fix this, we set user_agent for the session owned by
`Browser`, so that all requests use the specified user_agent.