Run the testsuite through pytest regardless of how it's called
The manual test infrastructure (with functions listed in the main
body of the file) has two major problems:

 1. It is incompatible with special pytest functionality, such
    as parametrize and fixture. These change the test function
    interface in a way that requires them to be run with pytest.

 2. The list of functions to manually test can get out of sync
    with those tested by pytest, since there is no way (other than
    manual inspection) to check that all the available tests are
    listed in the main body.

The test scripts can still be run manually as executables;
however, we switch from explicitly listing all functions to
collecting them automatically with pytest. This ensures that
the list stays in sync and that we can keep the pytest features
recently added to the test functions.

We pass `sys.argv` to `pytest.main` so that pytest options can
be used when the script is executed in this way (and because
`pytest.main` deprecated string arguments in pytest-3.0).