form.py: remove return value of choose_submit
Instead of a bool return value (True if the specified element was
found), raise a LinkNotFoundError exception if the element was not
found.

The default behavior should be safety, so we should not require
the user to verify that their specified submit element was found.
The exception can easily be caught in rare circumstances where
failure should not abort.

This change is in line with the error handling of other Form methods.