Recover executable functionality of test scripts
The test scripts can now be run manually as executables again;
however, we switch from explicitly listing all functions to
collecting them automatically with pytest. This ensures that
the list stays in sync and that we can keep the pytest features
recently added to the test functions.

We pass `sys.argv` to `pytest.main` so that pytest options can
be used when the script is executed in this way (and because
`pytest.main` deprecated string arguments in pytest-3.0).