use setup.py in travis.yml
We still need to install flake8 separately, but the tests really boils
down to "python setup.py test" now.

This avoids having to list dependencies twice. It means that the list of
dependencies that appear in setup.py are now tested: we'll notice if we
add anything and forget to modify setup.py.