Fix Travis-CI coverage report uploading
The `coverage` executable needs to exist in the environment for the
Codecov uploader to work correctly. Therefore, it is insufficient to
install `pytest-cov` (which provides `coverage`) with `tests_require`
in setup.py.

This fixes the "Python coveragepy not found" error output by the
uploader after the change to the Travis-CI environment in b3232dd.