tests: add project directory to sys.path
Fixes #100.

We had an inconsistency where "python setup.py test" was running the
in-project tests, but "python ./tests/test_xxx.py" was running the
installed version. This is error-prone, and in general at least I prefer
running tests against the non-installed version, so that debugging a
failed test can be done directly in-tree, without the need to re-install
anything.