README.md: more precise python version info
Use the PyPI badge to display which Python versions are supported in
the latest PyPI release version. Previously, Python version support
identified the version supported in the current master branch, but
these two are not necessarily the same.

Provide instructions for developers to see the Python version support
in the current master branch so that we do not need to hard-code the
version list in more places (it's already hard-coded in two places:
`setup.py` and `.travis.yml`).