Convert README.md to .rst format
Markdown is cool, but we're starting to use RST massively in the
documentation and it's painful to have to change the format when
moving something to or from README.

Also, Pypi likes RST, and we can upload our README.rst there once it's
in RST format.

This is the raw output from pandoc. A manual fixup follows.