Include tests directory in flake8 tests
This is mostly whitespace and line-length changes, as well as
a few unused variable fixes.

The tests `test_form_{noname,multiple}` now assert that the
request was successful.

The test `test_form_not_found` now uses the html example
"page_with_various_fields" because it was mostly a duplicate of
"page_with_missing_elements", which is now deleted.