.travis.yml: don't use flake8 in pypy environments
Closes #146.

Until we understand why flake8 is so slow in pypy environments
(to the point where it causes Travis to stall), do not run flake8
tests in these environments.

We do this by overriding the option specified in `setup.cfg` about
which files flake8 should ignore, telling it to instead ignore all
files. This is slightly hacky because a) we still want flake8 to
always be run by users and by all other environments and b) there
is no '--no-flake8' option.

We are still running flake8 tests in all standard python envs,
so there is no real loss of testing coverage.