Replace httpbin.org with pytest-httpbin in tests
The pytest-httpbin module provides pytest support for the
httpbin module (which is the code that runs the remote server
http://httpbin.org). This locally spins up an internal webserver
when tests are run.

With this change, MechanicalSoup tests can be run without an
internet connection. As a result, the tests run much faster.

You may need the python{,3}-dev package on your system to pip
install the pytest-httpbin module.