browser.py: refactor Browser.submit method
Closes #165.

Since both `Browser._prepare_request` and `Browser._build_request` are
internal methods, we can replace them with a single `Browser._request`.

This new method is the same as `_build_request`, except that it calls
`self.Session.request` to prepare and send the request instead of doing
it in two separate steps.

Since `requests.Session.request` both prepares and sends the request,
it has an API that includes keywords for both actions. The keywords
are a superset of `requests.Request` (except for `hooks`, but since
MechanicalSoup did not expose the resulting Request object, I believe
the hooks could not be used).

As a result, the `kwargs` that can be passed to `Browser.submit` has
expanded to include keywords related to sending the request.