Fixed a ReferenceError when using Python 3.4.3
This is a fix for ReferenceError: weakly-referenced object no longer
exists when mechanicalsoup is used on Python 3.4.3. self.session may
have already been garbage collected by the time browser.py gets deleted,
so calling self.session.close() will result in this error.

It's resolved by using weakref.finalize instead. It is safe to call
twice finalize on the same reference, so the call is just a
no-op (instead of the old error) if the object is already garbage
collected.

The tests and the FAQ are updated to reflect this change