form.py: harden choose_submit against invalid html
Closes #180.

If a submit element exists without a name-attribute, then previously
`Form.choose_submit` would raise a KeyError when it loops over the
elements.

Instead, we should simply ignore any nameless submit element,
regardless of whether it is there because the html is malformed or
because that particular element is used for javascript, for example.