Browser: make raise_on_404 public
This private __raise_on_404 is a bad habit from a C++/Java programmer.
We need to share this attribute with StatefulBrowser in the next
commit, and we don't want to go through the getters/setters pain. We
have Python properties in case we need to use getters/setters in the
future without breaking the API.