Provide a helpful error message when no form is selected
Since the currently selected form can now be accessed with
`StatefulBrowser.form`, it may not be obvious to users why this
returns `None` if they haven't called `StatefulBrowser.select_form`
yet.

Raising an exception in this case makes it unambiguous that
`form` in this context means "selected form", and that to use
this property they will need to first select a form.