browser.py: write temporary file with .html extension
Depending on the contents of the file (I think), Google Chrome may
display a file as text rather than html unless it has a .html
extension. In `launch_browser`, we typically want to look at the
rendered html of the page (otherwise we would just print the html
as text in the command line), so to ensure this happens we can
add a `.html` suffix to the temporary file.