test_form.py: fix test_choose_submit_twice
When a BeautifulSoup object is constructed from an html string,
it will standardize the html. In this case, since the input html
was just a raw <form>, it added <html> and <body> tags around it.

Since the `Form` ctor now demands that it is passed a <form> element,
we need to extract <form> from the object BeautifulSoup constructs
to pass to the `Form` ctor.