add new SVG logo
Fixes #111.

A previous attempt at a logo was made in #111, but the first logo was
only available as PNG. As discussed in #111 and on Gitter, add a more
profesional-looking logo, with SVG source and a PNG export for
inclusion in README.md (GitHub doesn't allow including SVG with
external fonts).

Include the logo in README.md and on the generated docs on
readthedocs.

Commit-message-by: Matthieu Moy <git@matthieu-moy.fr>
Original-patch-by: Mike Kormendy <info@mikekormendy.com>