Properly include the logo in the docs
Sphinx copies the image when using .. image:: ../assets/....png, so
there's no need to include the PNG from GitHub. Using the in-repo
logo, we're sure we generate a self-contained website for
documentation, and avoid hardcoding the repository name and branch in
the generated HTML.

While we're there, add an alt= text, center the logo, and put a bit of
space between the logo and section name.