form.py: relax error to deprecation warning
In 4e3fd3cdda9a, passing a bs4.element.Tag that was not a "form"
to the constructor of `Form` was escalated to a `LinkNotFoundError`.

Since this breaks backwards compatibility, we soften the error to
a deprecation warning (using the warnings library). In a future
version, this may become an error again.

However, `StatefulBrowser.select_form` has always raised an error
in this case since it began accepting a bs4.element.Tag in a5dec341,
so we continue to do so there.

Related to #228.