stateful_browser.py: rename submit_selected to submit
The obvious method to use when trying to submit a form is the `submit`
method. Unless you are following the documentation explicitly, you
may overlook the `submit_selected` method (which is really the only
correct method to use when submitting a form within a StatefulBrowser
instance).

This is demonstrated by #230, where the user got an unexpected error
because they used `submit` instead of `submit_selected`.

To adhere to the principle of least astonishment, make `submit` a
synonym for `submit_selected` (with `submit` being the primary name
and `submit_selected` being an alias for backwards compatibility).

This is a breaking change for anyone calling the base class `submit`
from a StatefulBrowser instance. It is arguable that even if the call
is made correctly, it is the "wrong" way to submit, so we should make
it more difficult to "do the wrong thing".

You can still access the base class `Browser.submit` from an instance
of `StatefulBrowser` if you need to by using the `super` builtin, e.g.

br = mechanicalsoup.StatefulBrowser()
super(StatefulBrowser, br).submit(...)