stateful_browser.py: improve submit backwards compatibility
In the (likely) most common scenario of an intended call to the
`Browser.submit` method from a StatefulBrowser instance, add a
backwards-compatible passthrough to `Browser.submit` in the new
`StatefulBrowser.submit` method while also issuing a deprecation
warning.

The way this works is that it checks to see if `btnName` is an
instance of the `Form` class, in which case it is unambiguously
intended to be a call to `Browser.submit`.

This should fix the majority of cases where users may have been
affected by the breaking change of overriding the `submit` method
in StatefulBrowser.

We cannot fix the remaining case, which is when `Browser.submit`
was passed a bs4.element.Tag with name attribute "form". The reason
we cannot create a similar passthrough for this is that there may
be a legitimate `btnName` that is a bs4.element.Tag with name="form".
This is unlikely (since submit elements typically would not be
named "form"), but is technically possible.