Improve `test_link_arg_*` tests
The `test_link_arg_regex` test was accidentally passing. By adding
another link to the test html, we can see that it was selecting the
first link instead of the link that matched the supplied regex.
This new version of the test fails, but it will be fixed by #256.

Also added a case for no arguments and reduced code duplication
by changing it to a parameterized test.