form.py: replace find_by_type with case-insensitive selectors
This change requires bs4 4.7.0+ (since older versions of bs4 do not
follow the CSS selector spec as closely as SoupSieve does).

I think we should delay applying this patch, since a hard dependency
on bs4 4.7.0 is too restrictive right now, but I do think that the
resulting code is more compact and easier to read.