Add "enctype respecting" feature and tests
Fixes #242.

Make Mechanical Soup behave as a real browser:
- only "multipart/form-data" enctype allow file sending
- forms without file inputs can now be sent as "multipart/form-data"
	if specified
- "application/x-www-form-urlencoded" is the default when enctype is wrong
	or not specified

The added code uses a hack because of a lack in Requests functionalities:
Requests doesn't support forcing enctype yet, and doesn't allow to submit
using "multipart/data-form" enctype without submitting files.
So the best way we found to bypass this is to pass an modified dict
as files to Requests.

This code could change in the future if Requests implements
a "force enctype" feature.