Improve handling of kwargs passed to Requests
We pass the form data to requests via `data` for POST methods, and
via `params` for GET methods. Therefore, we should be initializing
our data with the kwargs `data` or `params` respectively (whereas
previously we were initializing with `data` only).

Also added tests to check that we raise an exception when the kwargs
include a parameter that duplicates one of the args that we manually
set (i.e. `url` or `method`).