stateful_browser.py: add a response property
Whenever we set the browser state, include the request response.
The `response` property will raise an `AttributeError` if the state
has not yet been set.

This is not strictly necessary, since we return the request response
in most functions that have one, but it is perhaps a convenience in
some scenarios.

Closes #358.