Add GitHub Actions test workflow (#384)
Fixes #357.

The TravisCI gives a free plan to trigger tests in a limited way, that
is a bit restrictive for all the developers who want to contribute to
MechanicalSoup. A built-in CI in github called github actions is a good
solution to easily run workflows without paying credits.
We can create workflows by adding .yml files in the .github/workflows/
folder. This PR creates a file called python-package.yml that runs all
the different CI tests with the different python versions like with TravisCI.