Prevent test workflow from triggering twice
Restrict the branches on which the test workflow is triggered by pushes
so that we don't trigger it twice on PR branches that are created in
the upstream repository.

In general, most developers will create the PR branch in their fork,
but this is not the case for, e.g., PRs that dependabot creates.