Remove support for python 3.6 and 3.7
These versions reached end of life in Dec 2021 and Jun 2023
respectively. To address recent security vulnerabilities, we need to
update to newer versions of dependencies that no longer support these
EOL versions (e.g. #434, while this is a dev dependency issue only,
it's probably time to update the production requirements as well).